@@ -16,7 +16,7 @@ publish_to: none
 version: 1.9.3+221
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=3.0.0"
 
 dependencies:
   flutter:
@@ -25,7 +25,6 @@ dependencies:
     sdk: flutter
   dio: ^5.1.1
   crypto: ^3.0.2
-  sqflite: ^2.2.5
   path_provider: ^2.0.13
   shared_preferences: ^2.1.0
   share_plus: ^6.3.4
@@ -47,13 +46,16 @@ dependencies:
   flutter_colorpicker: ^1.0.3
   waterfall_flow: ^3.0.2
   animations: ^2.0.7
-  extended_text: ^10.0.1
+  extended_text: 
+    git: 
+      url: https://github.com/VB10/extended_text
+      ref: master
   font_awesome_flutter: ^10.1.0
   json_annotation: ^4.8.0
   flutter_staggered_animations: ^1.1.1
   badges: ^3.1.1
   image_picker: ^0.8.7+4
-  intl: ^0.17.0
+  intl: ^0.18.0
   quiver: ^3.2.1
   flutter_inappwebview_quill: ^5.4.6
   auto_animated: ^3.1.0
@@ -65,6 +67,26 @@ dependencies:
   easy_refresh: ^3.3.1+2
   flutter_hooks: ^0.18.6
   hooks_riverpod: ^2.3.6
+  sqflite: ^2.2.8+2
+  dynamic_color: ^1.6.3
+  # Design Style
+  fluent_ui: ^4.5.0
+  # Desktop Support
+  flutter_acrylic: ^1.1.1
+  window_manager: ^0.3.2
+  file_picker: ^5.2.11
+  # Windows Only
+  win32: ^3.1.4 # windows_single_instance: ^0.1.5 依赖但支持使用更高版本的 win32库
+  win32_registry: ^1.1.0
+  windows_single_instance: ^0.1.5
+  flutter_single_instance: ^0.0.1
+
+dependency_overrides:
+  win32: ^4.1.3
+  extended_text_library:
+    git: 
+      url: https://github.com/fluttercandies/extended_text_library
+      ref: master
 
 dev_dependencies:
   flutter_test:
@@ -73,6 +95,7 @@ dev_dependencies:
   mobx_codegen: ^2.1.1
   json_serializable: ^6.6.1
   analyzer: ^5.7.1
+  sqflite_common_ffi: ^2.2.5
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
