@@ -19,6 +19,8 @@ import 'dart:io';
 import 'package:bot_toast/bot_toast.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
+import 'package:pixez/constants.dart';
+import 'package:pixez/er/fluent_leader.dart';
 import 'package:pixez/er/lprinter.dart';
 import 'package:pixez/main.dart';
 import 'package:pixez/models/account.dart';
@@ -35,6 +37,10 @@ import 'package:url_launcher/url_launcher.dart';
 
 class Leader {
   static Future<void> pushUntilHome(BuildContext context) async {
+    if (Constants.isFluent) {
+      FluentLeader.pushUntilHome(context);
+      return;
+    }
     Navigator.of(context).pushAndRemoveUntil(
       MaterialPageRoute(
           builder: (context) =>
@@ -44,6 +50,10 @@ class Leader {
   }
 
   static Future<void> pushWithUri(BuildContext context, Uri link) async {
+    if (Constants.isFluent) {
+      FluentLeader.pushWithUri(context, link);
+      return;
+    }
     if (link.host == "script" && link.scheme == "pixez") {
       Navigator.of(context).push(MaterialPageRoute(builder: (context) {
         return SaveEvalPage(
@@ -266,14 +276,29 @@ class Leader {
     }
   }
 
-  static Future<dynamic> pushWithScaffold(context, Widget widget) {
+  static Future<dynamic> pushWithScaffold(context, Widget widget,
+      {Widget? icon, Widget? title}) {
+    if (Constants.isFluent) {
+      return FluentLeader.pushWithScaffold(context, widget,
+          icon: icon, title: title);
+    }
     return Navigator.of(context).push(MaterialPageRoute(
         builder: (context) => Scaffold(
               body: widget,
             )));
   }
 
-  static Future<dynamic> push(context, Widget widget) {
+  static Future<dynamic> push(
+    context,
+    Widget widget, {
+    Widget? icon,
+    Widget? title,
+    bool forceSkipWrap = false,
+  }) {
+    if (Constants.isFluent) {
+      return FluentLeader.push(context, widget,
+          icon: icon, title: title, forceSkipWrap: forceSkipWrap);
+    }
     return Navigator.of(context).push(MaterialPageRoute(
         builder: (context) => Scaffold(
               body: widget,
