@@ -14,7 +14,7 @@
  *
  */
 import 'package:path/path.dart';
-import 'package:sqflite/sqflite.dart';
+import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 
 class BanUserIdPersist {
   String? userId;
