@@ -30,20 +30,12 @@ import 'package:pixez/i18n.dart';
 import 'package:pixez/main.dart';
 import 'package:pixez/models/recommend.dart';
 import 'package:pixez/network/api_client.dart';
+import 'package:pixez/page/about/contributors.dart';
 import 'package:pixez/page/about/thanks_list.dart';
 import 'package:pixez/page/about/update_page.dart';
 import 'package:share_plus/share_plus.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:in_app_purchase_storekit/store_kit_wrappers.dart';
-
-class Contributor {
-  final String name;
-  final String avatar;
-  final String url;
-  final String content;
-
-  Contributor(this.name, this.avatar, this.url, this.content);
-}
 
 class AboutPage extends StatefulWidget {
   final bool? newVersion;
@@ -55,58 +47,6 @@ class AboutPage extends StatefulWidget {
 }
 
 class _AboutPageState extends State<AboutPage> {
-  List<Contributor> contributors = [
-    Contributor(
-        'Tragic Life',
-        'https://avatars3.githubusercontent.com/u/16817202?v=4',
-        'https://github.com/TragicLifeHu',
-        '🌍'),
-    Contributor(
-        'Skimige',
-        'https://avatars3.githubusercontent.com/u/9017470?v=4',
-        'https://xyx.moe/',
-        '📖'),
-    Contributor('Xian', 'https://avatars1.githubusercontent.com/u/34748039?v=4',
-        'https://github.com/itzXian', '🌍'),
-    Contributor(
-        'karin722',
-        'https://avatars0.githubusercontent.com/u/54385201?v=4',
-        'http://ivtune.net/',
-        '🌍'),
-    Contributor(
-        'Romani-Archman',
-        'https://avatars0.githubusercontent.com/u/68731023?v=4',
-        'http://archman.fun/',
-        '📖'),
-    Contributor(
-        'Henry-ZHR',
-        'https://avatars1.githubusercontent.com/u/51886614?s=64&v=4',
-        'https://github.com/Henry-ZHR',
-        '💻'),
-    Contributor(
-        'Takase',
-        'https://avatars0.githubusercontent.com/u/20792268?s=64&v=4',
-        'https://github.com/takase1121',
-        '🌍'),
-    Contributor(
-        'ChsBuffer',
-        'https://avatars3.githubusercontent.com/u/33744752?s=64&v=4',
-        'https://github.com/chsbuffer',
-        '💻'),
-    Contributor("媛天徵", "https://avatars.githubusercontent.com/u/64569368?v=4",
-        'https://github.com/YooTynChi', '🌍'),
-    Contributor(
-        "Scighost",
-        "https://avatars.githubusercontent.com/u/61003590?v=4",
-        'https://github.com/Scighost',
-        '💻'),
-    Contributor(
-        "sheason2019",
-        "https://avatars.githubusercontent.com/u/73812146?v=4",
-        'https://github.com/sheason2019',
-        '💻'),
-  ];
-
   late bool hasNewVersion;
   StreamSubscription<List<PurchaseDetails>>? _subscription;
   List<ProductDetails> products = [];
