@@ -15,501 +15,9 @@
  */
 
 import 'package:flutter/material.dart';
+import 'package:pixez/page/about/thanks_peoples.dart';
 
 class ThanksList extends StatelessWidget {
-  final peoples = [
-    "Leajen",
-    "欣殿",
-    "风间仓平",
-    "张孟航",
-    "给",
-    "傲天",
-    "折翼天使",
-    "不见",
-    "freecloud",
-    "炫酷的云宝",
-    "不会系鞋带的柴犬",
-    "得年",
-    "猫铳",
-    "Cyue",
-    "张维杰",
-    "清月",
-    "黑色阳光",
-    "DJMonet",
-    "怡桧",
-    "Fluency",
-    "轻风拂水光",
-    "千岛之间",
-    "夏",
-    "妄二郎",
-    "繁华—落幕",
-    "二十面体",
-    "手抖了?",
-    "VIP",
-    "通梓",
-    "大为",
-    "帅千",
-    "蕉迟但到",
-    "槿",
-    "桃花煮雨",
-    "毛巾",
-    "洋",
-    "吾姓朱",
-    "昕",
-    "他说他是",
-    "tra",
-    "雨嘉",
-    "○+●",
-    "浅梦灬深海",
-    "mxd",
-    "劲",
-    "龙战八荒",
-    "长歌行",
-    "墨黑",
-    "晨凯",
-    "空想 ",
-    "superstarcar",
-    "懿",
-    "鹏飞",
-    "六焰飞鱼",
-    "QAQ",
-    "夏天",
-    "Ag",
-    "野",
-    "小威",
-    "Mercis",
-    "LIMBO",
-    "VIG曦遥",
-    "撸猫",
-    "z7ui",
-    "*焯森",
-    "林",
-    "幽魂小瑾",
-    "Vincentzyx",
-    "WJM",
-    "奈",
-    "Artorius",
-    "moreki",
-    "布鲁斯赧",
-    "一只贱贱的熊",
-    "翱",
-    "DYCxhn",
-    "圈",
-    "indigle",
-    "m*6",
-    "*了",
-    "*蛙",
-    "NATSU",
-    "SunREVEN",
-    "Keter",
-    "程岚",
-    "略略略",
-    "志毅",
-    "墨烯",
-    "清DU欢",
-    "f*i",
-    "*永健",
-    "*恒",
-    "*泽",
-    "玄冰封月",
-    "*念",
-    "*宇",
-    "*外",
-    "*魔",
-    "*熙",
-    "*恒",
-    "*迹",
-    "路过的你",
-    "皮皮弛",
-    "*萍",
-    "?",
-    "韦涛涛",
-    "?",
-    "*文亮",
-    "不可思议的世界",
-    "*",
-    "l*c",
-    "D*h",
-    " *柄",
-    "s*r",
-    "o*n(lightinkqian)",
-    "天辰",
-    "青楼自管弦",
-    "不可思议的世界",
-    "*文亮",
-    "?",
-    "怎么就没电了",
-    "*色",
-    "*e",
-    "*喵",
-    "不屈意志",
-    "alchimie",
-    "なつき",
-    "灏",
-    "猛犸牙",
-    "404",
-    "横竖都是二",
-    "平 安",
-    "zwk",
-    "*",
-    "*子",
-    "*机",
-    "*环",
-    "浪*a",
-    "M*Q",
-    "智聪",
-    "6TBWhite",
-    " Stardust",
-    "哈哈",
-    "晏子敬",
-    "君不知",
-    "avileng",
-    "那你可zun棒哦",
-    "小杸",
-    "恩文",
-    "X*N",
-    "*吧",
-    "*弦",
-    "I*",
-    "*月",
-    "C*n",
-    "*军",
-    "*糕",
-    "*坑",
-    "*白",
-    "*梦",
-    "*罗",
-    "*格",
-    "*.",
-    "*蒂",
-    "S*H",
-    "*喵",
-    "*l",
-    "*莲",
-    "Nero",
-    "*琦公",
-    "明杰",
-    "*志强",
-    "勒紧裤腰",
-    "*泽鸿",
-    "bilibili村长",
-    "*承",
-    "*",
-    "*雷",
-    "d*t",
-    "*喵",
-    ":*|",
-    "辰",
-    "*子予",
-    "志鹏",
-    "2az",
-    "Myth",
-    "Namelost",
-    "韵秋",
-    "卷柏泡茶",
-    "闪耀的shining",
-    "相信自己",
-    "钰",
-    "*胜超",
-    "二货",
-    "*馮",
-    "*幕",
-    "*火",
-    "*】",
-    "如此精细",
-    "淼",
-    "百合菌",
-    "早睡早起",
-    "*延泽",
-    "光与暗的童话",
-    "Sheh",
-    "*铭越",
-    "相信自己",
-    "3m",
-    "烧仙草",
-    "*蒂",
-    "*5",
-    "*天",
-    "*水",
-    "*齐",
-    "*目",
-    "*剑",
-    "*晨",
-    "*傻",
-    "*桝",
-    "浩然",
-    "韦涛涛",
-    "Sana",
-    "云绛雨",
-    "*志远",
-    "守夜人",
-    "尧琦",
-    "q*e",
-    "*VII",
-    "*诺",
-    "*见",
-    "*少",
-    "*影",
-    "*子",
-    "*诗",
-    "*馨",
-    "无形丿无心",
-    "*玮峰",
-    "心缘",
-    "星叁夕",
-    "智儒",
-    "涛酱",
-    "丶式",
-    "Ludywus",
-    "小杸",
-    "岚思议",
-    "。。。",
-    "秋风轻拂",
-    "皓然",
-    "*侯",
-    "咻",
-    "*萤",
-    "*茨",
-    "*锐",
-    "M*e",
-    "*雪",
-    "*姐",
-    "*果",
-    "*鸟",
-    "*起",
-    "*眼",
-    "*阖",
-    "l*o",
-    "*』",
-    "Y*i",
-    "S*a",
-    "*恒",
-    "*生",
-    "*夏",
-    "*币",
-    "Q*n",
-    "*VII",
-    "*诺",
-    "*豆",
-    "*ル",
-    "*哥",
-    "*少",
-    "r*n",
-    "L*o",
-    "*思蒙",
-    "*萍",
-    "师匠",
-    "Master",
-    "一晨",
-    "钢筋混凝土",
-    "海山",
-    "KRISS",
-    "快,看天!",
-    "羽墨",
-    "京一",
-    "逍遥游",
-    "相信自己",
-    "MineLCF",
-    "啊八",
-    "YukiGulu",
-    "俊涛",
-    "南门陈",
-    "*佳凯",
-    "*兔",
-    "*g",
-    "*天",
-    "*人",
-    "*楼",
-    "*凡",
-    "*返",
-    "*阳",
-    "i*h",
-    "砂银",
-    "奥特Q",
-    "*乔辉",
-    "旭洋",
-    "xupite",
-    "家康",
-    "没有名字",
-    "苏端",
-    "世荣",
-    "陈浓药",
-    "伯霖",
-    "陈曦",
-    "天之所向",
-    "*放",
-    "浪",
-    "*辰",
-    "羽墨",
-    "逍遥游",
-    "MineLCF",
-    "啊八",
-    "YukiGulu",
-    "南门陈",
-    "俊涛",
-    "*佳凯",
-    "s*r",
-    "*商",
-    "*花",
-    "*梦",
-    "*率",
-    "*里",
-    "*鱼",
-    "*了",
-    "*r",
-    "o*d",
-    "*华",
-    "*锋",
-    "*g",
-    "*人",
-    "*雪",
-    "*月",
-    "?*?",
-    "*弦",
-    "*小",
-    "A*d",
-    "H*z",
-    "*得",
-    "*!",
-    "*根",
-    "*z",
-    "*川",
-    "R*n",
-    "*易",
-    "*急",
-    "*飞",
-    "*兔",
-    "*楼",
-    "*凡",
-    "*皇",
-    "Y*x",
-    "N*I",
-    "*了",
-    "*蹊",
-    "*孩",
-    "M*D",
-    "E*L",
-    "f*n",
-    "*文",
-    "*香",
-    "z*r",
-    "*年",
-    "H*A",
-    "*卿",
-    "T*e",
-    "F*e",
-    "*饼",
-    "S*s",
-    "*威",
-    "*诺",
-    "*日",
-    "*月",
-    "*龙",
-    "*超",
-    "M*i",
-    "*云",
-    "J*m",
-    "W*T",
-    "s*r",
-    "T*k",
-    "z*i",
-    "H*A",
-    "nolanly",
-    "命运石之门",
-    "tb",
-    "杰",
-    "伟泽",
-    "浅笑",
-    "齐",
-    "仙",
-    "弘博",
-    "ravioli",
-    "空灵幽魂",
-    "风流云散",
-    "吞噬椰菜花的跳蚤",
-    "胖胖的兔子",
-    "EventuallyFade",
-    "相信自己",
-    "3yxt",
-    "邪王真眼",
-    "曹森麒",
-    "Erico",
-    "zjc",
-    "🌅",
-    "SquarePants",
-    "20.2",
-    "刘海下的金色瞳",
-    "90s",
-    "HHHHYଓ",
-    "靠近",
-    "fool",
-    "Powaaaaaildi",
-    "淦",
-    "松竹子",
-    "Hsk",
-    "中华摸鱼协会首席技术顾问",
-    "Claybot.aic",
-    "pretty boi",
-    "名字什么的不重要",
-    "乐乐",
-    "F",
-    "毓",
-    "Susan.",
-    "yoink",
-    "三星堆青铜人面像",
-    "爽",
-    "厘米",
-    "禾",
-    "ゆずりは いのり",
-    "过零丁杨",
-    "玻璃伞",
-    "wjl",
-    "怀牧",
-    "TL-TCT",
-    "无恒",
-    "....",
-    "zjc",
-    "静水",
-    "皇帝不急太监急",
-    "天晓得",
-    "你猜",
-    "永远的盾拉杰",
-    "W s",
-    "虎啸天穹",
-    "隐者哒",
-    "潇潇雨辰",
-    "松鼠",
-    "打你家玻璃",
-    "陈年凉夜忆星",
-    "松竹子",
-    "移风易俗",
-    "雾雨幽森",
-    "鹏飞",
-    "HK",
-    "国祥",
-    "Wallpaper",
-    "画风",
-    "风流云散",
-    "就随便中文",
-    "翀",
-    "丽超",
-    "洁",
-    "绯月琉海",
-    "r",
-    "风逸飞",
-    "clear_wind",
-    "LouZT 06 ",
-    "卡咖酷",
-    "格克琳尔",
-    "绮夜",
-    "QAQ",
-    "清静",
-    "SF",
-    "无尽的回忆",
-    "黑的焰火",
-    "三石"
-  ];
-
   @override
   Widget build(BuildContext context) {
     return ListView(
