@@ -141,39 +141,37 @@ class _AndroidHelloPageState extends State<AndroidHelloPage> {
                   duration: const Duration(milliseconds: 400),
                   height: isFullscreen ? 0 : bottomNavigatorHeight,
                   key: bottomNavigatorKey,
-                  child: BottomNavigationBar(
-                      type: BottomNavigationBarType.fixed,
-                      selectedItemColor: Theme.of(context).colorScheme.primary,
-                      currentIndex: index,
-                      onTap: (index) {
-                        if (this.index == index) {
-                          topStore.setTop("${index + 1}00");
-                        }
-                        setState(() {
-                          this.index = index;
-                        });
-                        if (_pageController.hasClients)
-                          _pageController.jumpToPage(index);
-                      },
-                      items: [
-                        BottomNavigationBarItem(
-                            icon: Icon(Icons.home),
-                            label: I18n.of(context).home),
-                        BottomNavigationBarItem(
-                            icon: Icon(
-                              CustomIcons.leaderboard,
-                            ),
-                            label: I18n.of(context).rank),
-                        BottomNavigationBarItem(
-                            icon: Icon(Icons.favorite),
-                            label: I18n.of(context).quick_view),
-                        BottomNavigationBarItem(
-                            icon: Icon(Icons.search),
-                            label: I18n.of(context).search),
-                        BottomNavigationBarItem(
-                            icon: Icon(Icons.more_horiz),
-                            label: I18n.of(context).more),
-                      ]),
+                  child: NavigationBar(
+                    destinations: [
+                      NavigationDestination(
+                          icon: Icon(Icons.home), label: I18n.of(context).home),
+                      NavigationDestination(
+                          icon: Icon(
+                            Icons.leaderboard,
+                          ),
+                          label: I18n.of(context).rank),
+                      NavigationDestination(
+                          icon: Icon(Icons.favorite),
+                          label: I18n.of(context).quick_view),
+                      NavigationDestination(
+                          icon: Icon(Icons.search),
+                          label: I18n.of(context).search),
+                      NavigationDestination(
+                          icon: Icon(Icons.more_horiz),
+                          label: I18n.of(context).more)
+                    ],
+                    selectedIndex: index,
+                    onDestinationSelected: (index) {
+                      if (this.index == index) {
+                        topStore.setTop("${index + 1}00");
+                      }
+                      setState(() {
+                        this.index = index;
+                      });
+                      if (_pageController.hasClients)
+                        _pageController.jumpToPage(index);
+                    },
+                  ),
                 )));
   }
 
